fix(macos): harden notification session cleanup and error reporting
